Stored View base class in util directory
