+++ fitnessmonitor/views.py
@@ -1,19 +1,7 @@
 import tkinter as tk
 import tkinter.ttk as ttk
 
-
-class View(ttk.Frame):
-    view_list = {}
-    def __init__(self, name: str, parent: tk.Tk, controller):
-        super().__init__(parent)
-        self._parent = parent
-        self._controller = controller
-        self.view_list[name] = self
-        self.grid(row=0, column=0, sticky="nsew")
-    
-    def _initWidgets(self, title):
-        ttk.Label(self, text=title, font=("Helvetica", 32), anchor="center").grid(row=0, column=0, sticky=tk.NSEW)
-  
+from util.view import View
         
 class MainView(View):
     def __init__(self, parent: tk.Tk, controller):
